Move test that is interfering with other tests to own file
