+++ src/App/tests/App.test.tsx
@@ -1,4 +1,3 @@
-import * as sessionDux from '@onaio/session-reducer';
 import { mount } from 'enzyme';
 import toJson from 'enzyme-to-json';
 import { createBrowserHistory } from 'history';
@@ -87,43 +86,6 @@ describe('App', () => {
     expect(supersetUserIsLogged).toBeTruthy();
     expect(supersetUserIsUser).toBeTruthy();
     wrapper.unmount();
-  });
-
-  it('attempts to logout user', async () => {
-    fetch.mockResponse(JSON.stringify(expressAPIResponse));
-    delete window.location;
-    const hrefMock = jest.fn();
-    (window.location as any) = {
-      set href(url: string) {
-        hrefMock(url);
-      },
-    };
-
-    const mockLogout = jest.fn(() => null);
-    const wrapper = mount(
-      <Provider store={store}>
-        <Router history={history}>
-          <App logoutComponent={mockLogout} />
-        </Router>
-      </Provider>
-    );
-    await new Promise<unknown>(resolve => setImmediate(resolve));
-    wrapper.update();
-
-    // At this point we have an authenticated user
-    const loggedIn = sessionDux.isAuthenticated(store.getState());
-    expect(loggedIn).toBeTruthy();
-
-    // simulate logout
-    history.push('/logout');
-    wrapper.update();
-
-    // unfortunately we don't have a definitive way to test that the user session was invalidated;
-    // since the functionality that does this is in the express server and is thus out of the
-    // react-app's scope.
-    // ensure we are calling the logout component which handles opensrp logout and redirects to backend
-    // PS: only after redirecting to the express server and back is the user unauthenticated.
-    expect(mockLogout).toBeCalled();
   });
 
   it('tracks dimensions correctly', async () => {
